pr/vkmbox/82: revert unpermitted changes
