+++ .gitignore
@@ -8,7 +8,6 @@ __pycache__/
 
 # Distribution / packaging
 .Python
-dp/deepy/
 build/
 develop-eggs/
 dist/
