+++ docker-compose.yml
@@ -23,7 +23,7 @@ services:
         dff-funfact-skill:8104, dff-bot-persona-skill:8105, news-api-annotator:8112,
         dff-gossip-skill:8109, dff-wiki-skill:8111, dff-gaming-skill:8115, topic-recommendation:8113,
         user-persona-extractor:8114, wiki-facts:8116, dff-music-skill:8099, entity-detection:8103, dff-art-skill:8117,
-        dff-template-skill:8120, dff-intent-responder-skill:8121"
+        dff-template-skill:8120"
       WAIT_HOSTS_TIMEOUT: ${WAIT_TIMEOUT:-480}
     deploy:
       resources:
