--- skills/dff_intent_responder_skill/requirements.txt
@@ -1,2 +1,0 @@
-click==7.1.2
-nltk==3.5
