--- skills/dff_intent_responder_skill/test.sh
@@ -1,3 +1,0 @@
-#!/bin/bash
-
-python test_server.py
