+++ skills/dff_template_skill/scenario/condition.py
@@ -1,15 +1,18 @@
 import logging
 
-from df_engine.core import Context, Actor
-
-from common.dff.integration import condition as int_cnd
+from df_engine.core import Actor, Context
+from scenario.response_funcs import get_respond_funcs
+import common.utils as common_utils
+import common.dff.integration.context as int_ctx
 
 logger = logging.getLogger(__name__)
-# ....
 
 
-def example_lets_talk_about():
-    def example_lets_talk_about_handler(ctx: Context, actor: Actor, *args, **kwargs) -> str:
-        return int_cnd.is_lets_chat_about_topic_human_initiative(ctx, actor)
+def intent_catcher_exists_condition(ctx: Context, actor: Actor, *args, **kwargs) -> bool:
 
-    return example_lets_talk_about_handler
+    intents_by_catcher = common_utils.get_intents(
+        int_ctx.get_last_human_utterance(ctx, actor), probs=False, which="intent_catcher"
+    )
+
+    response_funcs = get_respond_funcs()
+    return bool(any([intent in response_funcs for intent in intents_by_catcher]))
