--- skills/dff_template_skill/scenario/processing.py
@@ -1,5 +1,0 @@
-import logging
-
-
-logger = logging.getLogger(__name__)
-# ....
