--- skills/dff_template_skill/tests/lets_talk_out.json
@@ -1,142 +1,0 @@
-[
-    [
-        [
-            "Hi, how are you?",
-            "Hi, what's up?"
-        ],
-        [
-            1.0,
-            1.0
-        ],
-        [
-            {
-                "dff_template_skill_state": {
-                    "shared_memory": {},
-                    "previous_human_utter_index": 0,
-                    "history": {
-                        "0": [
-                            "greeting",
-                            "node1"
-                        ]
-                    },
-                    "current_turn_dff_suspended": false,
-                    "context": {
-                        "id": "15b477a2-562a-416f-b1ea-057e5722bb4f",
-                        "labels": {
-                            "0": [
-                                "greeting",
-                                "node1"
-                            ]
-                        },
-                        "requests": {
-                            "0": "hi."
-                        },
-                        "responses": {
-                            "0": [
-                                [
-                                    "Hi, how are you?",
-                                    0.0,
-                                    {},
-                                    {},
-                                    {}
-                                ],
-                                [
-                                    "Hi, what's up?",
-                                    0.0,
-                                    {},
-                                    {},
-                                    {}
-                                ]
-                            ]
-                        },
-                        "misc": {
-                            "slots": {
-                                "topic": "science",
-                                "user_name": "Gordon Freeman"
-                            }
-                        },
-                        "validation": false,
-                        "actor_state": {}
-                    }
-                },
-                "dff_shared_state": {
-                    "cross_states": {},
-                    "cross_links": {}
-                },
-                "used_links": {},
-                "age_group": "unknown",
-                "disliked_skills": []
-            },
-            {
-                "dff_template_skill_state": {
-                    "shared_memory": {},
-                    "previous_human_utter_index": 0,
-                    "history": {
-                        "0": [
-                            "greeting",
-                            "node1"
-                        ]
-                    },
-                    "current_turn_dff_suspended": false,
-                    "context": {
-                        "id": "15b477a2-562a-416f-b1ea-057e5722bb4f",
-                        "labels": {
-                            "0": [
-                                "greeting",
-                                "node1"
-                            ]
-                        },
-                        "requests": {
-                            "0": "hi."
-                        },
-                        "responses": {
-                            "0": [
-                                [
-                                    "Hi, how are you?",
-                                    0.0,
-                                    {},
-                                    {},
-                                    {}
-                                ],
-                                [
-                                    "Hi, what's up?",
-                                    0.0,
-                                    {},
-                                    {},
-                                    {}
-                                ]
-                            ]
-                        },
-                        "misc": {
-                            "slots": {
-                                "topic": "science",
-                                "user_name": "Gordon Freeman"
-                            }
-                        },
-                        "validation": false,
-                        "actor_state": {}
-                    }
-                },
-                "dff_shared_state": {
-                    "cross_states": {},
-                    "cross_links": {}
-                },
-                "used_links": {},
-                "age_group": "unknown",
-                "disliked_skills": []
-            }
-        ],
-        [
-            {},
-            {}
-        ],
-        [
-            {
-                "can_continue": "can"
-            },
-            {
-                "can_continue": "can"
-            }
-        ]
-    ]
-]
